requirements.txt
MetaTrader5==5.0.4424
pandas==2.2.2
numpy==1.26.4
python-telegram-bot==21.0.1
# ... other dependencies ...